+++ GalaxyZooTouchTable/GalaxyZooTouchTable/GalaxyZooTouchTable.csproj
@@ -158,7 +158,7 @@
     <Compile Include="Views\Centerpiece.xaml.cs">
       <DependentUpon>Centerpiece.xaml</DependentUpon>
     </Compile>
-    <Compile Include="ClassificationSummary.xaml.cs">
+    <Compile Include="Views\ClassificationSummary.xaml.cs">
       <DependentUpon>ClassificationSummary.xaml</DependentUpon>
     </Compile>
     <Compile Include="Config.cs" />
@@ -179,7 +179,7 @@
     <Compile Include="Views\StartButton.xaml.cs">
       <DependentUpon>StartButton.xaml</DependentUpon>
     </Compile>
-    <Compile Include="SubjectViewer.xaml.cs">
+    <Compile Include="Views\SubjectViewer.xaml.cs">
       <DependentUpon>SubjectViewer.xaml</DependentUpon>
     </Compile>
     <Compile Include="Views\UserConsole.xaml.cs">
@@ -190,7 +190,6 @@
     <Compile Include="ViewModels\ClassificationPanelViewModel.cs" />
     <Compile Include="ViewModels\ExamplesPanelViewModel.cs" />
     <Compile Include="ViewModels\MainWindowViewModel.cs" />
-    <Compile Include="ViewModels\SubjectViewerViewModel.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ClassificationData.cs" />
@@ -264,7 +263,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="ClassificationSummary.xaml">
+    <Page Include="Views\ClassificationSummary.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -308,7 +307,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="SubjectViewer.xaml">
+    <Page Include="Views\SubjectViewer.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
