+++ GalaxyZooTouchTable/GalaxyZooTouchTable/Models/AnswerButton.cs
@@ -11,7 +11,6 @@ namespace GalaxyZooTouchTable.Models
         public string Url { get; set; }
         public int Index { get; set; }
         public int AnswerCount { get; set; } = 0;
-        public bool SelectedAnswer { get; set; } = false;
 
         /// <summary>
         /// The Galaxy Zoo workflow contains image links in Markdown.
