--- GalaxyZooTouchTable/GalaxyZooTouchTable/ViewModels/SubjectViewerViewModel.cs
@@ -1,15 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace GalaxyZooTouchTable.ViewModels
-{
-    public class SubjectViewerViewModel : INotifyPropertyChanged
-    {
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        private void OnPropertyRaised(string propertyname)
-        {
-            if (PropertyChanged != null)
-                PropertyChanged(this, new PropertyChangedEventArgs(propertyname));
-        }
-    }
-}
