+++ GalaxyZooTouchTable/GalaxyZooTouchTable/Views/ClassificationSummary.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace GalaxyZooTouchTable
+{
+    /// <summary>
+    /// Interaction logic for ClassificationSummary.xaml
+    /// </summary>
+    public partial class ClassificationSummary : UserControl
+    {
+        public ClassificationSummary()
+        {
+            InitializeComponent();
+        }
+    }
+}
