+++ GalaxyZooTouchTable/GalaxyZooTouchTable/Views/SubjectViewer.xaml
