+++ GalaxyZooTouchTable/GalaxyZooTouchTable/Views/SubjectViewer.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace GalaxyZooTouchTable
+{
+    /// <summary>
+    /// Interaction logic for SubjectViewer.xaml
+    /// </summary>
+    public partial class SubjectViewer : UserControl
+    {
+        public SubjectViewer()
+        {
+            InitializeComponent();
+        }
+    }
+}
